Struct soa::soa2::Soa2

pub struct Soa2<A, B> {
    // some fields omitted
}

A growable struct-of-2-arrays type, with heap allocated contents.

This structure is analogous to a Vec<(A, B)>, but instead of laying out the tuples sequentially in memory, each row gets its own allocation. For example, an Soa2<f32, i64> will contain two inner arrays: one of f32s, and one of i64s.

All data is aligned to 16-bytes. Feel free to do SIMD operations with array contents.

Methods

impl<A, B> Soa2<A, B>

fn new() -> Soa2<A, B>

Constructs a new, empty Soa2.

The SoA will not allocate until elements are pushed onto it.

fn with_capacity(capacity: usize) -> Soa2<A, B>

Constructs a new, empty Soa2 with the specified capacity.

The SoA will be able to hold exactly capacity tuples of elements without reallocating.

If capacity is 0, the SoA will not allocate.

It is important to note that this function does not specify the length of the soa, but only the capacity.

unsafe fn from_raw_parts(ptra: *mut A, ptrb: *mut B, len: usize, cap: usize) -> Soa2<A, B>

Constructs a Soa2 directly from the raw components of another.

This is highly unsafe, and no invariants are checked.

unsafe fn from_raw_bufs(ptra: *const A, ptrb: *const B, elts: usize) -> Soa2<A, B>

Constructs a Soa2 by copying the elements from raw pointers.

This function will copy elts contiguous elements from each of the pointers into a new allocation owned by the returned Soa2. The elements of the buffer are copied without cloning, as if ptr::read() were called on them.

fn from_vecs(v0: Vec<A>, v1: Vec<B>) -> Soa2<A, B>

Constructs a Soa2 directly from vectors of its components.

This function will panic if the lengths of the vectors don't match.

If the capacity of the vectors don't match they will be reallocated to have matching capacities.

Otherwise, no allocation will be performed and the SoA will only take ownership of the elements in the vectors.

fn len(&self) -> usize

Returns the number of tuples stored in the SoA.

fn is_empty(&self) -> bool

Returns true if the SoA contains no elements.

unsafe fn set_len(&mut self, len: usize)

Sets the length of a vector.

This will explicitly set the size of the soa, without actually modifying its buffers, so it is up to the caller to ensure that the SoA is actually the specified size.

fn capacity(&self) -> usize

Returns the number of elements the SoA can hold without reallocating.

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given SoA. The collection may reserve more space to avoid frequent reallocations.

Panics if the new capacity overflows usize.

fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given SoA. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Panics if the new capacity overflows usize.

fn shrink_to_fit(&mut self)

Shrinks the capacity of the SoA as much as possible.

It will drop down as close as possible to the length, but the allocator may still inform the SoA that there is space for a few more elements.

fn truncate(&mut self, len: usize)

Shorten a SoA, dropping excess elements.

If len is greater than the soa's current length, this has no effect.

fn as_mut_slices<'a>(&'a mut self) -> (&'a mut [A], &'a mut [B])

Returns mutable slices over the SoA's elements.

fn as_slices<'a>(&'a self) -> (&'a [A], &'a [B])

Returns slices over the SoA's elements.

fn iters(&self) -> (Iter<A>, Iter<B>)

Returns iterators over the SoA's elements.

fn zip_iter(&self) -> Zip<Iter<A>, Iter<B>>

Returns a single iterator over the SoA's elements, zipped up.

fn iters_mut(&mut self) -> (IterMut<A>, IterMut<B>)

Returns mutable iterators over the SoA's elements.

fn zip_iter_mut(&mut self) -> Zip<IterMut<A>, IterMut<B>>

Returns a single iterator over the SoA's elements, zipped up.

fn into_iters(self) -> (IntoIter<A>, IntoIter<B>)

Converts an SoA into iterators for each of its arrays.

fn into_vecs(self) -> (Vec<A>, Vec<B>)

Converts an SoA into a pair of Vecs. This will neither allocator nor copy.

fn as_ptrs(&self) -> (*const A, *const B)

Returns a pair of pointers to the start of the data in an SoA.

fn as_mut_ptrs(&mut self) -> (*mut A, *mut B)

Returns a pair of pointers to the start of the mutable data in an SoA.

fn swap_remove(&mut self, index: usize) -> (A, B)

Removes an element from anywhere in the SoA and returns it, replacing it with the last element.

This does not preserve ordering, but is O(1).

Panics if index is out of bounds.

fn insert(&mut self, index: usize, element: (A, B))

Inserts an element at position index within the vector, shifting all elements after position index one position to the right.

Panics if index is not between 0 and the SoA's length, inclusive.

fn remove(&mut self, index: usize) -> (A, B)

Removes and returns the elements at position index within the SoA, shifting all elements after position index one position to the left.

Panics if index is out of bounds.

fn retain<F>(&mut self, f: F) where F: FnMut((&A, &B)) -> bool

Returns only the element specified by the predicate.

In other words, remove all elements e such that f(&e) returns false. This method operates in place and preserves the order of the retained elements.

fn push(&mut self, value: (A, B))

Appends an element to the back of a collection.

Panics if the number of elements in the SoA overflows a usize.

fn pop(&mut self) -> Option<(A, B)>

Removes the last element from a SoA and returns it, or None if empty.

fn append(&mut self, other: &mut Self)

Moves all the elements of other into self, leaving other empty.

Panics if the number of elements in the SoA overflows a usize.

fn clear(&mut self)

Clears the SoA, removing all values.

fn extend<I0, I1>(&mut self, i0: I0, i1: I1) where I0: Iterator<Item=A>, I1: Iterator<Item=B>

Extends the SoA with the elements yielded by arbitrary iterators.

Panics (and leaks memory!) if the iterators yield a different number of elements.

fn from_iters<I0, I1>(i0: I0, i1: I1) -> Soa2<A, B> where I0: Iterator<Item=A>, I1: Iterator<Item=B>

Constructs an Soa2 with elements yielded by arbitrary iterators.

Panics (and leaks memory!) if the iterators yield a different number of elements.

impl<A: Clone, B: Clone> Soa2<A, B>

fn resize(&mut self, new_len: usize, value: (A, B))

Resizes the SoA in-place so that len() is equal to new_len.

Calls either extend() or truncate() depending on whether new_len is larger than the current value of len() or not.

fn push_all(&mut self, x0: &[A], x1: &[B])

Appends all elements in slices to the SoA.

Iterates over the slices, clones each element, and then appends them to this SoA. The slices are traversed one at a time, in order.

Panics if the slices are of different lengths.